refactor(ponder): 重构 Ponder 场景代码结构
- 修改了 Ponder 场景的包结构，按功能分类存放场景类
- 更新了 AnvilCraftPonderScenes 类中的导入路径和注册顺序
- 重命名了部分场景类所在的包：
  - HeaterScene 移动到 recipe 包
  - ImpactPileScene 移动到 structure 包
