+++ src/main/java/dev/dubhe/anvilcraft/integration/ponder/AnvilCraftPonderScenes.java
@@ -2,11 +2,11 @@ package dev.dubhe.anvilcraft.integration.ponder;
 
 import dev.dubhe.anvilcraft.integration.ponder.scene.AnvilScene;
 import dev.dubhe.anvilcraft.integration.ponder.scene.power.TransmissionPoleScene;
-import dev.dubhe.anvilcraft.integration.ponder.scene.HeaterScene;
-import dev.dubhe.anvilcraft.integration.ponder.scene.ImpactPileScene;
+import dev.dubhe.anvilcraft.integration.ponder.scene.recipe.HeaterScene;
+import dev.dubhe.anvilcraft.integration.ponder.scene.recipe.SpaceOvercompressorScene;
 import dev.dubhe.anvilcraft.integration.ponder.scene.redstone.BlockComparatorScene;
 import dev.dubhe.anvilcraft.integration.ponder.scene.redstone.MagnetScene;
-import dev.dubhe.anvilcraft.integration.ponder.scene.recipe.SpaceOvercompressorScene;
+import dev.dubhe.anvilcraft.integration.ponder.scene.structure.ImpactPileScene;
 import net.createmod.ponder.api.registration.PonderSceneRegistrationHelper;
 import net.minecraft.resources.ResourceLocation;
 
@@ -14,6 +14,8 @@ public class AnvilCraftPonderScenes {
     public static void register(PonderSceneRegistrationHelper<ResourceLocation> helper) {
         // base
         AnvilScene.register(helper);
+        // power
+        TransmissionPoleScene.register(helper);
         // recipe
         SpaceOvercompressorScene.register(helper);
         HeaterScene.register(helper);
@@ -22,7 +24,5 @@ public class AnvilCraftPonderScenes {
         BlockComparatorScene.register(helper);
         // structure
         ImpactPileScene.register(helper);
-        // power
-        TransmissionPoleScene.register(helper);
     }
 }
