+++ src/main/java/dev/dubhe/anvilcraft/integration/ponder/scene/structure/ImpactPileScene.java
@@ -1,4 +1,4 @@
-package dev.dubhe.anvilcraft.integration.ponder.scene;
+package dev.dubhe.anvilcraft.integration.ponder.scene.structure;
 
 
 import dev.dubhe.anvilcraft.init.block.ModBlocks;
